+++ src/main/java/io/github/manuelarte/spring/queryparameter/config/QueryCriteriaConfig.java
@@ -9,12 +9,12 @@ import java.util.List;
 import org.springframework.beans.factory.annotation.Qualifier;
 import org.springframework.boot.autoconfigure.condition.ConditionalOnMissingBean;
 import org.springframework.context.annotation.Bean;
-import org.springframework.context.annotation.ComponentScan;
 import org.springframework.context.annotation.Configuration;
+import org.springframework.context.annotation.Import;
 import org.springframework.core.convert.ConversionService;
 
 @Configuration
-@ComponentScan("io.github.manuelarte.spring.queryparameter.operators")
+@Import(OperatorConfig.class)
 @lombok.RequiredArgsConstructor
 public class QueryCriteriaConfig {
 
@@ -28,7 +28,7 @@ public class QueryCriteriaConfig {
 
   @Bean("defaultTypeTransformer")
   public TypeTransformer<?, ?> defaultTypeTransformer(final ConversionService conversionService) {
-    return new ClassFieldTransformerImpl(conversionService);
+    return new ClassFieldTransformerImpl<>(conversionService);
   }
 
   @Bean