+++ src/main/java/io/github/manuelarte/spring/queryparameter/operators/EqualsOperator.java
@@ -1,14 +1,7 @@
 package io.github.manuelarte.spring.queryparameter.operators;
 
-import org.springframework.core.annotation.Order;
-import org.springframework.stereotype.Component;
-
-@Component
-@Order(EqualsOperator.ORDER)
-@lombok.EqualsAndHashCode
+@lombok.EqualsAndHashCode(callSuper = true)
 public class EqualsOperator extends AbstractMiddleOperator {
-
-  public static final int ORDER = 0;
 
   private static final String OPERATOR = "::";