+++ src/main/java/io/github/manuelarte/spring/queryparameter/operators/GreaterThanOperator.java
@@ -1,14 +1,7 @@
 package io.github.manuelarte.spring.queryparameter.operators;
 
-import org.springframework.core.annotation.Order;
-import org.springframework.stereotype.Component;
-
-@Component
-@Order(GreaterThanOperator.ORDER)
-@lombok.EqualsAndHashCode
+@lombok.EqualsAndHashCode(callSuper = true)
 public class GreaterThanOperator extends AbstractMiddleOperator {
-
-  public static final int ORDER = GreaterThanOrEqualsOperator.ORDER + 10;
 
   private static final String OPERATOR = ":>";